+++ src/components/CurrentSection/CurrentSection.d.ts
@@ -1,0 +1,4 @@
+export interface CurrentSectionProps {
+	title: string;
+	desc: string;
+}