+++ src/components/EncourageMessage/index.tsx
@@ -1,10 +1,11 @@
 import { FC } from 'react';
+import { EncourageMessageProps } from './EncourageMessage.d';
 import './EncourageMessage.scss';
 
-export const EncourageMessage: FC = () => {
+export const EncourageMessage: FC<EncourageMessageProps> = ({ message }) => {
 	return (
 		<div className="box box-white encourage-message">
-			<p>100% ти можеш більше! Ми в тебе віримо! Ти - молодець!</p>
+			<p>{message}</p>
 		</div>
 	);
 };